Add difference calculation
difference_data contains data created by the script Calculate_Difference.ipynb. In this file, the different metrics are compared between the data of the authors and our preprocessed data. The analysis reveals discrepancies in the metrics across the two datasets, indicating variations in the data generation process or potential issues with data integrity or transformation.